all library objects are now referred to with their absolute paths
These libraries constitute cross-directory communication in the make. As such,
the prerequisites and targets of pairs of recipes must match EXACTLY in order to
be matched up. For instance, if a rule calls for libRetro/libRetro.so and
another rule says how build libAttic/../libRetro/libRetro.so, make gets
confused. Using absolute paths resolves this