I now clean out all the .o files explicitly by name.
This is better than the wildcard because it deals with objects in all the
directories. For instance, if I had

LIB_OBJECTS = a/b.o

then the previous method of cleaning out *.o would not work, but the new method
does. I also still clean out *.o in case there are any other object files
floating around. These should have been defined in EXTRACLEAN, but I do this
anyway for now