Headers are now installed into a directory based on the header location in the tree
This is for compatibility with the g-speak tree

the headers are installed into
  /usr/include/$(HEADER_PATH)/$(HEADER_SUBPATH)

Here the HEADER_PATH holds all of this project's headers, and if the
project has sub-path headers, they go into subdirectories. The
HEADER_PATH defaults to the location of the project in the source
tree, unless explicitly overwridden in the project Makefile. Example:
project lives in $(ROOT)/libXXX. It's trying to install a.h, d/b.h.
These will be installed to
  /usr/include/libXXX/a.h
  /usr/include/libXXX/d/b.h