fixed improper handling of $(dir)_CHILDREN_LDLIBS_PROJECT
I had a rule to build statically-linked executables:

exe: exe.o $(LIB_TARGET_A) $(LDLIBS_PROJECT:%.$(SO)=%.a) $($(dir)_CHILDREN_LDLIBS_PROJECT)
     recipe

As always, the targets and prerequisites are evaluated immediatly. This
was problematic because $(dir)_CHILDREN_LDLIBS_PROJECT is built up as
the project Makefiles are scanned. Thus when this rule is parsed,
$(dir)_CHILDREN_LDLIBS_PROJECT may not yet have its final value. This
patch removes the dependency, and processes the variable in the RECIPE,
which is evaluated deferred-ly. The dependency was added to the .a file
rule. Thus now all the children .a files must be built for a parent .a
to be considered done