All objects going into DSOs now built with -fPIC
Suppose I have a.c that's to be a part of libA.so. Ideally, I'd build it
with -fPIC when it's going to libA.so, and without -fPIC when it's going
to the statically-linked libA.a. Thus in the ideal case, 'make install'
would compile a.c twice, since 'make install' should build both flavors
of libraries.

Before this patch, I was doing this as described, by compiling a.c into
two different object files: a.o and a-fPIC.o. This worked fine, EXCEPT
when the user specified per-target options, things went awry. In a
project Makefile the user can say

$(dir)/a.o: CCXXFLAGS += -Dasdf

In the scheme I had, this flag would ONLY apply to the no-fPIC version,
which is a bug. Some ideas I have is to ALWAYS compile everything twice,
so do something like

%.o %-fPIC.o: %.c
    gcc -o $(^:.c=.o) ....
    gcc -o $(^:.c=-fPIC.o) -fPIC ....

The down side is that build times are now ALWAYS twice as long. Another
option is to use a recursive make to build the other flavor of object.

I'm choosing to ignore this issue for now, and to apply -fPIC to all
objects going into a DSO. This would incur a performance penalty for all
the statically-linked uses, but it's probably negligible.